Recursive merge of sub messages when duplicated on the wire
Excerpt from https://developers.google.com/protocol-buffers/docs/encoding#optional:

Normally, an encoded message would never have more than one instance of a non-repeated field. However, parsers are expected to handle the case in which they do.

For numeric types and strings, if the same field appears multiple times, the parser accepts the last value it sees.  (Note: already handled by protox)

For embedded message fields, the parser merges multiple instances of the same field, as if with the Message::MergeFrom method – that is, all singular scalar fields in the latter instance replace those in the former, singular embedded messages are merged, and repeated fields are concatenated. The effect of these rules is that parsing the concatenation of two encoded messages produces exactly the same result as if you had parsed the two messages separately and merged the resulting objects.