Don't sort fields for encoding
https://developers.google.com/protocol-buffers/docs/encoding#order now states:
"When a message is serialized, there is no guaranteed order for how its known or unknown fields should be written. Serialization order is an implementation detail and the details of any particular implementation may change in the future. Therefore, protocol buffer parsers must be able to parse fields in any order."