Always compile well-known types
It's necessary to do so otherwise the compilation of proto definitions that don't include these well-known types will fail.
Indeed, the JSON decoding/encoding modules need the modules generated for the well-known-types to be compiled.

Still need to fix the warnings caused by this: now, modules of well-known types may be generated more that once.